Clean up Python and CPP editors
