--- client/app/bundles/course/assessment/question/programming/components/OnlineEditor/Cpp/OnlineEditorCppView.intl.js
@@ -1,9 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  appendSubtitle: {
-    id: 'course.assessment.question.programming.onlineEditorCppView.appendSubtitle',
-    defaultMessage: 'For any function/variable declarations',
-    description: 'Subtitle for append code block.',
-  },
-});
