--- client/app/bundles/course/assessment/question/programming/constants/programmingQuestionConstants.jsx
@@ -1,17 +1,0 @@
-import mirrorCreator from 'mirror-creator';
-
-const actionTypes = mirrorCreator([
-  'PROGRAMMING_QUESTION_UPDATE',
-  'SKILLS_UPDATE',
-  'EDITOR_MODE_UPDATE',
-  'HAS_ERROR_CLEAR',
-  'SUBMISSION_MESSAGE_SET',
-  'SUBMISSION_MESSAGE_CLEAR',
-  'SUBMIT_FORM_LOADING',
-  'SUBMIT_FORM_SUCCESS',
-  'SUBMIT_FORM_FAILURE',
-  'SUBMIT_FORM_EVALUATING',
-  'VALIDATION_ERRORS_SET',
-]);
-
-export default actionTypes;
